verified_requirement.txt
#python3.12
accelerate==1.10.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
bidict==0.23.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
controlnet-aux==0.0.3
cycler==0.12.1
diffusers==0.27.2
easydict==1.13
einops==0.8.1
fastapi==0.108.0
ffmpy==0.6.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
gradio==4.21.0
gradio_client==0.12.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.24.6
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
IOPaint==1.4.3
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.0.1
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.11.1
packaging==25.0
pandas==2.3.1
peft==0.7.1
piexif==1.1.3
Pillow==9.5.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.7.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.33
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
ruff==0.12.8
safetensors==0.6.2
scikit-image==0.24.0
scipy==1.16.1
semantic-version==2.10.0
setuptools==78.1.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.32.0.post1
sympy==1.14.0
tifffile==2025.6.11
timm==1.0.19
tokenizers==0.21.4
tomlkit==0.12.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.48.3
triton==3.4.0
typer==0.16.0
typer-config==1.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==11.0.3
wheel==0.45.1
wsproto==1.2.0
yacs==0.1.8
zipp==3.23.0
